+++ client/src/app/modules/image-viewers/widgets/context-image/image-options/image-options.component.ts
@@ -37,11 +37,12 @@ import { RangeSliderValue } from "src/app/modules/pixlisecore/components/atoms/r
 import { APICachedDataService } from "src/app/modules/pixlisecore/services/apicacheddata.service";
 import { ImageListReq, ImageListResp } from "src/app/generated-protos/image-msgs";
 import { ScanImagePurpose } from "src/app/generated-protos/image";
-import { APIDataService, SnackbarService } from "src/app/modules/pixlisecore/pixlisecore.module";
+import { AnalysisLayoutService, APIDataService, ContextImageDataService, SnackbarService } from "src/app/modules/pixlisecore/pixlisecore.module";
 import { ImportMarsViewerImageReq, ImportMarsViewerImageResp } from "src/app/generated-protos/image-coreg-msgs";
 import { MinMax } from "src/app/models/BasicTypes";
 import { ImageBeamLocationVersionsReq, ImageBeamLocationVersionsResp } from "src/app/generated-protos/image-beam-location-msgs";
 import { ScanListReq, ScanListResp } from "src/app/generated-protos/scan-msgs";
+import { WidgetType } from "../../../../widget/models/widgets.model";
 
 export class ImageDisplayOptions {
   constructor(
@@ -83,6 +84,7 @@ export class ImageDisplayOptions {
 }
 export class ImagePickerParams {
   constructor(
+    public widgetId: string,
     public scanIds: string[],
     public warningMsg: string,
     public options: ImageDisplayOptions
@@ -127,16 +129,23 @@ export class ImageOptionsComponent implements OnInit, OnDestroy {
 
   loadingBeamVersions: boolean = false;
 
+  private _targetWidgetIds: string[] = [];
+
+  widgetTypes: WidgetType[] = ['context-image'];
+
   constructor(
+    private _analysisLayoutService: AnalysisLayoutService,
     private _cachedDataService: APICachedDataService,
     private _dataService: APIDataService,
     private _snackService: SnackbarService,
+    private _contextImageDataService: ContextImageDataService,
     @Inject(MAT_DIALOG_DATA) public data: ImagePickerParams,
     public dialogRef: MatDialogRef<ContextImagePickerComponent, ImagePickerResult>,
     public dialog: MatDialog //private _exportDataService: ExportDataService
   ) {
     // Copy the options so we can have "reset" buttons for eg
     this.loadOptions(data.options);
+    this._targetWidgetIds = [data.widgetId];
   }
 
   loadOptions(options: ImageDisplayOptions) {
@@ -175,10 +184,12 @@ export class ImageOptionsComponent implements OnInit, OnDestroy {
   }
 
   ngOnDestroy() {
+    this._analysisLayoutService.highlightedWidgetIds$.next([]);
     this._subs.unsubscribe();
   }
 
   onClose(): void {
+    this._analysisLayoutService.highlightedWidgetIds$.next([]);
     this.dialogRef.close();
   }
 
@@ -248,6 +259,9 @@ export class ImageOptionsComponent implements OnInit, OnDestroy {
   // Notifying caller of our options changing
   private publishOptionChange() {
     this.optionChange.emit(this.makeImagePickerResult());
+    for (const widgetId of this._targetWidgetIds) {
+      this._contextImageDataService.setWidgetImagePickerResult(widgetId, this.makeImagePickerResult());
+    }
   }
 
   private makeImagePickerResult(): ImagePickerResult {
@@ -628,6 +642,20 @@ export class ImageOptionsComponent implements OnInit, OnDestroy {
     //     console.error(`Error exporting images: ${err}`);
     //   }
     // );
+  }
+
+  get activeWidgetIds(): string[] {
+    return this._targetWidgetIds;
+  }
+
+  set activeWidgetIds(ids: string[]) {
+    this._targetWidgetIds = ids;
+    this._analysisLayoutService.highlightedWidgetIds$.next(this._targetWidgetIds);
+  }
+
+  onActiveWidgetIdsChanged(ids: string[]) {
+    this.activeWidgetIds = ids;
+    this.publishOptionChange();
   }
 
   onImport() {